% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{cormat}
\alias{cormat}
\title{Matrix pairwise correlation}
\usage{
cormat(x, diag = FALSE, upper = FALSE, missingThresh = 0.5)
}
\arguments{
\item{x}{numeric matrix where rows are variables to compare (e.g. genes) and
columns are conditions or samples.}

\item{diag}{logical to compute the diagonal elements.}

\item{upper}{logical to show upper part of the matrix.}

\item{missingThresh}{fraction of paired data that is required for the
correlation to be calculated.}
}
\value{
An R dist object containing the Pearson correlation for each
}
\description{
This function will take a matrix from R and do all pairwise
correlation calculations for each row in the matrix and
return a distance object containing the Pearson correlations.

The values returned should be equivalent to running
cor(x,y,method="pearson",use="pair") where x and y are the
data from two rows being compared.

In addition, if the fraction of missing pair data between
the two matrix rows is greater than the missingThresh parameter,
the correlation is not calculated and an NA value is returned.

2006/02/22 (Mike Schaffer)
}
\examples{

# compute pairwise Pearson correlations across rows
cormat(rbind(c(2.5,4.2,1.6,6.2,3.1), c(5.2,3.5,8.2,7.4,5.3), c(0.3, 0.4, 1.4, 2.7, 3.1) ))

# compare cormat output to base R cor output:
cormat(rbind(c(1,1,3,4,NA),c(1,2,3,4,5) ) )  # 0.9467293
cor(c(1,1,3,4,NA),c(1,2,3,4,5),use="pair")   # 0.9467293

}
