% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\name{get.all.dists}
\alias{get.all.dists}
\title{Extract values from dist object for a given element}
\usage{
get.all.dists(row, dis)
}
\arguments{
\item{row}{index of row in the dist object}

\item{dis}{dist object}
}
\value{
vector of distance values
}
\description{
Convenience function to extract distances (or correlations) from a dist object
}
\examples{
corrs<-cormat(rbind(c(2.5,4.2,1.6,6.2,3.1), c(5.2,3.5,8.2,7.4,5.3), c(0.3, 0.4, 1.4, 2.7, 3.1) ))
get.all.dists(1, corrs)

}
